Based on feedback from Carlo - Stack based agent/states now have the ability to do previous actions before change state on stack..
States now have a "re-enter" method allowing agents to for example, return to their last position after seeking something. The state manager handles the entering, I also implemented a timer so when re-entering it stays in that state for a specified time before being able to change into another state.